requirements.txt
beautifulsoup4==4.11.1
binance==0.3
candlestick==0.0.8
candlestick_patterns_subodh101==1.1.0
kaleido
pandas==1.5.1
pandas_ta==0.3.14b0
plotly==5.10.0
python_binance==1.0.16
python_telegram_bot==13.13
requests==2.28.1
tweepy==3.10.0
pytest==7.2.0
yfinance
streamlit
duckdb-engine